Move 'invalid lookup' test to class creation tests
